requirements.txt
tensorflow>=2.13,<2.19
numpy>=1.24,<2.1
pandas>=2.0,<2.3
scikit-learn>=1.3,<1.7
matplotlib>=3.7,<3.11
seaborn>=0.13,<0.14
Pillow>=10.0,<12.0
opencv-python>=4.8,<4.12
fastapi>=0.111,<0.116
uvicorn[standard]>=0.30,<0.35
python-multipart>=0.0.9,<0.0.11
